Rename import module
