+++ flake-info/src/data/mod.rs
@@ -1,10 +1,10 @@
-mod derivation;
+mod import;
 mod export;
 mod flake;
 mod source;
 mod system;
 
-pub use derivation::{Derivation, Kind};
+pub use import::{Derivation, Kind};
 pub use export::Export;
 pub use flake::{Flake, Repo};
 pub use source::{FlakeRef, Hash, Source};
